requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
bcrypt
python-multipart
email-validator